LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-stepMismatch-expected.txt


PASS [INPUT in DATE status] The step attribute is not set 
PASS [INPUT in DATE status] The value attibute is empty string 
PASS [INPUT in DATE status] The value must match the step 
PASS [INPUT in DATE status] The value must mismatch the step 
PASS [INPUT in MONTH status] The step attribute is not set 
PASS [INPUT in MONTH status] The value attibute is empty string 
PASS [INPUT in MONTH status] The value must match the step 
PASS [INPUT in MONTH status] The value must mismatch the step 
PASS [INPUT in WEEK status] The step attribute is not set 
PASS [INPUT in WEEK status] The value attibute is empty string 
PASS [INPUT in WEEK status] The value must match the step 
PASS [INPUT in WEEK status] The value must mismatch the step 
PASS [INPUT in TIME status] The step attribute is not set 
PASS [INPUT in TIME status] The value attibute is empty string 
PASS [INPUT in TIME status] The value must match the step 
PASS [INPUT in TIME status] The value must mismatch the step 
PASS [INPUT in DATETIME-LOCAL status] The step attribute is not set 
PASS [INPUT in DATETIME-LOCAL status] The value attibute is empty string 
PASS [INPUT in DATETIME-LOCAL status] The value must match the step 
PASS [INPUT in DATETIME-LOCAL status] The value must mismatch the step 
PASS [INPUT in NUMBER status] The step attribute is not set 
PASS [INPUT in NUMBER status] The value attibute is empty string 
PASS [INPUT in NUMBER status] The value must match the step 
PASS [INPUT in NUMBER status] The value must mismatch the step 
